+++ src/ProjectMarworyn/Generators/DiceGenerator.cs
@@ -7,6 +7,13 @@
             return new Random(worldSeed);
         }
 
+        public int Next(Random random,
+            int startInclusive,
+            int endExclusive)
+        {
+            return random.Next(startInclusive, endExclusive);
+        }
+
         public double NextDouble(Random random)
         {
             return random.NextDouble();