+++ src/ProjectMarworyn/Models/Person.cs
@@ -7,6 +7,7 @@ namespace ProjectMarworyn.Models
         public int Id { get; set; }
         public Name Name { get; set; }
         public int Age { get; set; }
+        public Biosex Biosex { get; set; }
         public Gender Gender { get; set; }
         public bool WillHaveChildren { get; set; }
         public bool IsAlive { get; set; }