+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ---
 
-## [1.0.2] - In Progress
+## [1.1.0] - In Progress
 
 -Fixed issue where no one could die
 -Claude PRs now apply to all branches, not just `main` and `Update/**`
@@ -10,6 +10,8 @@
 -Name.json is now InitialPeople.json
 +As a side effect of the above: DeathEngine tuple has been removed
 -AIs are now prohibited from git push or git commit
+-Added Insex people - They currently don't pair
+-Seperated gender and biosex - Children are now generated from biosex pairs
 
 ## [1.0.1] - 2nd May 2026